feat: rename login page component
